GitButler Integration Commit
This is an integration commit for the virtual branches that GitButler is tracking.

Due to GitButler managing multiple virtual branches, you cannot switch back and
forth between git branches and virtual branches easily. 

If you switch to another branch, GitButler will need to be reinitialized.
If you commit on this branch, GitButler will throw it away.

Here are the branches that are currently applied:
 - misc (refs/gitbutler/misc)
   branch head: 37488820ccf205623b54a2e20e3a04e166e11c03
 - changes.
 (refs/gitbutler/changes.-)
   branch head: 85981c8d1b3c33a2fdffb3a46972c08a0d655197
 - misc 1 (refs/gitbutler/misc-1)
   - backend/src/graphql/movies/customMoviesResolver.ts
 - misc (refs/gitbutler/misc)
   branch head: 37488820ccf205623b54a2e20e3a04e166e11c03
For more information about what we're doing here, check out our docs:
https://docs.gitbutler.com/features/virtual-branches/integration-branch